fix: admin check for deleting users

+++ ayunis-core-backend/src/iam/users/application/use-cases/delete-user/delete-user.use-case.ts
@@ -1,6 +1,8 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { ForbiddenException, Injectable, Logger } from '@nestjs/common';
 import { UsersRepository } from '../../ports/users.repository';
 import { DeleteUserCommand } from './delete-user.command';
+import { UserNotFoundError, UserUnauthorizedError } from '../../users.errors';
+import { UserRole } from 'src/iam/users/domain/value-objects/role.object';
 
 @Injectable()
 export class DeleteUserUseCase {
@@ -10,6 +12,21 @@ export class DeleteUserUseCase {
 
   async execute(command: DeleteUserCommand): Promise<void> {
     this.logger.log('deleteUser', { userId: command.userId });
+    const requestUser = await this.usersRepository.findOneById(
+      command.requestUserId,
+    );
+    if (!requestUser) {
+      throw new UserNotFoundError(command.requestUserId);
+    }
+    if (
+      requestUser.orgId !== command.orgId ||
+      requestUser.role !== UserRole.ADMIN
+    ) {
+      throw new UserUnauthorizedError(
+        'You are not allowed to delete this user',
+      );
+    }
+
     await this.usersRepository.delete(command.userId);
   }
 }
